+++ packages/backend/src/controllers/v2/QueryController.ts
@@ -27,7 +27,6 @@ import {
 import {
     Body,
     Get,
-    Hidden,
     Middlewares,
     OperationId,
     Path,
@@ -73,7 +72,7 @@ export class QueryController extends BaseController {
         const results = await this.services
             .getAsyncQueryService()
             .getAsyncQueryResults({
-                user: req.user!,
+                account: req.account!,
                 projectUuid,
                 queryUuid,
                 page,
@@ -98,7 +97,7 @@ export class QueryController extends BaseController {
         this.setStatus(200);
 
         await this.services.getAsyncQueryService().cancelAsyncQuery({
-            user: req.user!,
+            account: req.account!,
             projectUuid,
             queryUuid,
         });
@@ -139,7 +138,7 @@ export class QueryController extends BaseController {
         const results = await this.services
             .getAsyncQueryService()
             .executeAsyncMetricQuery({
-                user: req.user!,
+                account: req.account!,
                 projectUuid,
                 invalidateCache: body.invalidateCache,
                 metricQuery,
@@ -171,7 +170,7 @@ export class QueryController extends BaseController {
         const results = await this.services
             .getAsyncQueryService()
             .executeAsyncSavedChartQuery({
-                user: req.user!,
+                account: req.account!,
                 projectUuid,
                 invalidateCache: body.invalidateCache,
                 chartUuid: body.chartUuid,
@@ -204,7 +203,7 @@ export class QueryController extends BaseController {
         const results = await this.services
             .getAsyncQueryService()
             .executeAsyncDashboardChartQuery({
-                user: req.user!,
+                account: req.account!,
                 projectUuid,
                 invalidateCache: body.invalidateCache,
                 chartUuid: body.chartUuid,
@@ -240,7 +239,7 @@ export class QueryController extends BaseController {
         const results = await this.services
             .getAsyncQueryService()
             .executeAsyncUnderlyingDataQuery({
-                user: req.user!,
+                account: req.account!,
                 projectUuid,
                 invalidateCache: body.invalidateCache,
                 underlyingDataSourceQueryUuid:
@@ -275,7 +274,7 @@ export class QueryController extends BaseController {
         const results = await this.services
             .getAsyncQueryService()
             .executeAsyncSqlQuery({
-                user: req.user!,
+                account: req.account!,
                 projectUuid,
                 invalidateCache: body.invalidateCache ?? false,
                 sql: body.sql,
@@ -307,7 +306,7 @@ export class QueryController extends BaseController {
         const results = await this.services
             .getAsyncQueryService()
             .executeAsyncSqlChartQuery({
-                user: req.user!,
+                account: req.account!,
                 projectUuid,
                 invalidateCache: body.invalidateCache ?? false,
                 context: context ?? QueryExecutionContext.SQL_RUNNER,
@@ -340,7 +339,7 @@ export class QueryController extends BaseController {
         const results = await this.services
             .getAsyncQueryService()
             .executeAsyncDashboardSqlChartQuery({
-                user: req.user!,
+                account: req.account!,
                 projectUuid,
                 invalidateCache: body.invalidateCache ?? false,
                 dashboardUuid: body.dashboardUuid,
@@ -380,7 +379,7 @@ export class QueryController extends BaseController {
         const readStream = await this.services
             .getAsyncQueryService()
             .getResultsStream({
-                user: req.user!,
+                account: req.account!,
                 projectUuid,
                 queryUuid,
             });
@@ -416,7 +415,7 @@ export class QueryController extends BaseController {
         this.setStatus(200);
 
         const results = await this.services.getAsyncQueryService().download({
-            user: req.user!,
+            account: req.account!,
             projectUuid,
             queryUuid,
             type: body.type,